Split out tokenizer
