+++ src/femr/models/tokenizer/__init__.py
@@ -1,0 +1,1 @@
+from .hierarchical_tokenizer import HierarchicalTokenizer
